+++ framework/src/mfem/equation_systems/EquationSystem.C
@@ -265,16 +265,9 @@ EquationSystem::FormSystem(mfem::OperatorHandle & op,
 }
 
 void
-EquationSystem::AssembleJacobian(
-    Moose::MFEM::NamedFieldsMap<mfem::ParBilinearForm> & jac_blfs,
-    Moose::MFEM::NamedFieldsMap<Moose::MFEM::NamedFieldsMap<mfem::ParMixedBilinearForm>> &
-        jac_mblfs,
-    Moose::MFEM::NamedFieldsMap<mfem::ParLinearForm> & rhs_lfs,
-    std::vector<mfem::Array<int>> & ess_tdof_lists,
-    std::vector<std::unique_ptr<mfem::ParGridFunction>> & var_ess_constraints,
-    mfem::OperatorHandle & op,
-    mfem::BlockVector & trueX,
-    mfem::BlockVector & trueRHS)
+EquationSystem::FormLegacySystem(mfem::OperatorHandle & op,
+                                 mfem::BlockVector & trueX,
+                                 mfem::BlockVector & trueRHS)
 {
   // Allocate block operator
   DeleteAllBlocks();
@@ -283,12 +276,12 @@ EquationSystem::AssembleJacobian(
   for (const auto i : index_range(_test_var_names))
   {
     auto & test_var_name = _test_var_names.at(i);
-    auto blf = jac_blfs.Get(test_var_name);
-    auto lf = rhs_lfs.Get(test_var_name);
+    auto blf = _blfs.Get(test_var_name);
+    auto lf = _lfs.Get(test_var_name);
     mfem::Vector aux_x, aux_rhs;
     mfem::HypreParMatrix * aux_a = new mfem::HypreParMatrix;
     blf->FormLinearSystem(
-        ess_tdof_lists.at(i), *(var_ess_constraints.at(i)), *lf, *aux_a, aux_x, aux_rhs);
+        _ess_tdof_lists.at(i), *(_var_ess_constraints.at(i)), *lf, *aux_a, aux_x, aux_rhs);
     _h_blocks(i, i) = aux_a;
     trueX.GetBlock(i) = aux_x;
     trueRHS.GetBlock(i) = aux_rhs;
@@ -305,13 +298,13 @@ EquationSystem::AssembleJacobian(
       mfem::Vector aux_x, aux_rhs;
       mfem::ParLinearForm aux_lf(_test_pfespaces.at(i));
       aux_lf = 0.0;
-      if (jac_mblfs.Has(test_var_name) && jac_mblfs.Get(test_var_name)->Has(trial_var_name))
+      if (_mblfs.Has(test_var_name) && _mblfs.Get(test_var_name)->Has(trial_var_name))
       {
-        auto mblf = jac_mblfs.Get(test_var_name)->Get(trial_var_name);
+        auto mblf = _mblfs.Get(test_var_name)->Get(trial_var_name);
         mfem::HypreParMatrix * aux_a = new mfem::HypreParMatrix;
-        mblf->FormRectangularLinearSystem(ess_tdof_lists.at(j),
-                                          ess_tdof_lists.at(i),
-                                          *(var_ess_constraints.at(j)),
+        mblf->FormRectangularLinearSystem(_ess_tdof_lists.at(j),
+                                          _ess_tdof_lists.at(i),
+                                          *(_var_ess_constraints.at(j)),
                                           aux_lf,
                                           *aux_a,
                                           aux_x,
@@ -327,14 +320,6 @@ EquationSystem::AssembleJacobian(
 
   // Create monolithic matrix
   op.Reset(mfem::HypreParMatrixFromBlocks(_h_blocks));
-}
-
-void
-EquationSystem::FormLegacySystem(mfem::OperatorHandle & op,
-                                 mfem::BlockVector & trueX,
-                                 mfem::BlockVector & trueRHS)
-{
-  AssembleJacobian(_blfs, _mblfs, _lfs, _ess_tdof_lists, _var_ess_constraints, op, trueX, trueRHS);
 }
 
 void
@@ -467,228 +452,6 @@ EquationSystem::BuildEquationSystem()
   BuildBilinearForms();
   BuildMixedBilinearForms();
   BuildLinearForms();
-}
-
-TimeDependentEquationSystem::TimeDependentEquationSystem(
-    const Moose::MFEM::TimeDerivativeMap & time_derivative_map)
-  : _dt_coef(1.0), _time_derivative_map(time_derivative_map)
-{
-}
-
-void
-TimeDependentEquationSystem::SetTimeStep(mfem::real_t dt)
-{
-  if (fabs(dt - _dt_coef.constant) > 1.0e-12 * dt)
-  {
-    _dt_coef.constant = dt;
-    for (auto test_var_name : _test_var_names)
-    {
-      auto blf = _blfs.Get(test_var_name);
-      blf->Update();
-      blf->Assemble();
-    }
-  }
-}
-
-void
-TimeDependentEquationSystem::SetTrialVariableNames()
-{
-  EquationSystem::SetTrialVariableNames();
-  // Add old state of trial variable as an eliminated variable
-  for (const auto & test_var_name : _test_var_names)
-  {
-    if (!VectorContainsName(_eliminated_var_names, test_var_name))
-      _eliminated_var_names.push_back(test_var_name);
-  }
-}
-
-void
-TimeDependentEquationSystem::AddKernel(std::shared_ptr<MFEMKernel> kernel)
-{
-  const auto & coupled_var_name = kernel->getTrialVariableName();
-  const auto & test_var_name = kernel->getTestVariableName();
-
-  if (_time_derivative_map.isTimeDerivative(coupled_var_name))
-  {
-    AddTestVariableNameIfMissing(test_var_name);
-    AddCoupledVariableNameIfMissing(coupled_var_name);
-    const auto & coupled_var_time_integral_name =
-        _time_derivative_map.getTimeIntegralName(coupled_var_name);
-    // Register new kernels map if not present for the test variable
-    if (!_td_kernels_map.Has(test_var_name))
-    {
-      auto kernel_field_map =
-          std::make_shared<Moose::MFEM::NamedFieldsMap<std::vector<std::shared_ptr<MFEMKernel>>>>();
-      _td_kernels_map.Register(test_var_name, std::move(kernel_field_map));
-    }
-    if (!_td_kernels_map.Get(test_var_name)->Has(coupled_var_time_integral_name))
-    {
-      auto kernels = std::make_shared<std::vector<std::shared_ptr<MFEMKernel>>>();
-      _td_kernels_map.Get(test_var_name)
-          ->Register(coupled_var_time_integral_name, std::move(kernels));
-    }
-    _td_kernels_map.GetRef(test_var_name)
-        .Get(coupled_var_time_integral_name)
-        ->push_back(std::move(kernel));
-  }
-  else
-  {
-    EquationSystem::AddKernel(kernel);
-  }
-}
-
-void
-TimeDependentEquationSystem::BuildBilinearForms()
-{
-  // Register bilinear forms
-  for (const auto i : index_range(_test_var_names))
-  {
-    auto test_var_name = _test_var_names.at(i);
-
-    // Apply kernels to blf
-    _blfs.Register(test_var_name, std::make_shared<mfem::ParBilinearForm>(_test_pfespaces.at(i)));
-    auto blf = _blfs.GetShared(test_var_name);
-    blf->SetAssemblyLevel(_assembly_level);
-    ApplyBoundaryBLFIntegrators<mfem::ParBilinearForm>(
-        test_var_name, test_var_name, blf, _integrated_bc_map, _dt_coef.constant);
-    ApplyDomainBLFIntegrators<mfem::ParBilinearForm>(
-        test_var_name, test_var_name, blf, _kernels_map, _dt_coef.constant);
-    // Apply dt*du/dt contributions from the operator on the trial variable
-    ApplyDomainBLFIntegrators<mfem::ParBilinearForm>(
-        test_var_name, test_var_name, blf, _td_kernels_map);
-    // Assemble
-    blf->Assemble();
-
-    // Apply kernels to td_blf
-    _td_blfs.Register(test_var_name,
-                      std::make_shared<mfem::ParBilinearForm>(_test_pfespaces.at(i)));
-    auto td_blf = _td_blfs.GetShared(test_var_name);
-    td_blf->SetAssemblyLevel(_assembly_level);
-    // ApplyBoundaryBLFIntegrators<mfem::ParBilinearForm>(
-    //     test_var_name, test_var_name, td_blf, _integrated_bc_map);
-    ApplyDomainBLFIntegrators<mfem::ParBilinearForm>(
-        test_var_name, test_var_name, td_blf, _td_kernels_map);
-    // Assemble
-    td_blf->Assemble();
-  }
-}
-
-void
-TimeDependentEquationSystem::BuildMixedBilinearForms()
-{
-  // Register mixed bilinear forms. Note that not all combinations may
-  // have a kernel.
-
-  // Create mblf for each test/coupled variable pair with an added kernel.
-  // Mixed bilinear forms with coupled variables that are not trial variables are
-  // associated with contributions from eliminated variables.
-  for (const auto i : index_range(_test_var_names))
-  {
-    auto test_var_name = _test_var_names.at(i);
-    auto test_mblfs = std::make_shared<Moose::MFEM::NamedFieldsMap<mfem::ParMixedBilinearForm>>();
-    for (const auto j : index_range(_coupled_var_names))
-    {
-      const auto & coupled_var_name = _coupled_var_names.at(j);
-      auto mblf = std::make_shared<mfem::ParMixedBilinearForm>(_coupled_pfespaces.at(j),
-                                                               _test_pfespaces.at(i));
-      // Register MixedBilinearForm if kernels exist for it, and assemble
-      // kernels
-      if (test_var_name != coupled_var_name)
-      {
-        // Apply all mixed kernels with this test/trial pair
-        ApplyBoundaryBLFIntegrators<mfem::ParMixedBilinearForm>(
-            coupled_var_name, test_var_name, mblf, _integrated_bc_map, _dt_coef.constant);
-        ApplyDomainBLFIntegrators<mfem::ParMixedBilinearForm>(
-            coupled_var_name, test_var_name, mblf, _kernels_map, _dt_coef.constant);
-        // Apply dt*du/dt contributions from the operator on the trial variable
-        ApplyDomainBLFIntegrators<mfem::ParMixedBilinearForm>(
-            coupled_var_name, test_var_name, mblf, _td_kernels_map);
-        if (mblf->GetDBFI()->Size() || mblf->GetBBFI()->Size())
-        {
-          // Assemble mixed bilinear forms
-          mblf->SetAssemblyLevel(_assembly_level);
-          mblf->Assemble();
-          // Register mixed bilinear forms associated with a single trial variable
-          // for the current test variable
-          test_mblfs->Register(coupled_var_name, mblf);
-        }
-      }
-    }
-    // Register all mixed bilinear form sets associated with a single test
-    // variable
-    _mblfs.Register(test_var_name, test_mblfs);
-  }
-
-  // Register mixed bilinear forms. Note that not all combinations may
-  // have a kernel.
-
-  // Create mblf for each test/trial variable pair with an added kernel
-  for (const auto i : index_range(_test_var_names))
-  {
-    const auto & test_var_name = _test_var_names.at(i);
-    auto test_td_mblfs =
-        std::make_shared<Moose::MFEM::NamedFieldsMap<mfem::ParMixedBilinearForm>>();
-    for (const auto j : index_range(_trial_var_names))
-    {
-      const auto & trial_var_name = _trial_var_names.at(j);
-      auto td_mblf = std::make_shared<mfem::ParMixedBilinearForm>(_test_pfespaces.at(j),
-                                                                  _test_pfespaces.at(i));
-      // Register MixedBilinearForm if kernels exist for it, and assemble
-      // kernels
-      if (test_var_name != trial_var_name)
-      {
-        // Apply all mixed kernels with this test/trial pair
-        ApplyDomainBLFIntegrators<mfem::ParMixedBilinearForm>(
-            trial_var_name, test_var_name, td_mblf, _td_kernels_map);
-        // Assemble mixed bilinear form
-        if (td_mblf->GetDBFI()->Size() || td_mblf->GetBBFI()->Size())
-        {
-          td_mblf->SetAssemblyLevel(_assembly_level);
-          td_mblf->Assemble();
-          // Register mixed bilinear forms associated with a single trial variable
-          // for the current test variable
-          test_td_mblfs->Register(trial_var_name, td_mblf);
-        }
-      }
-    }
-    // Register all mixed bilinear forms associated with a single test variable
-    _td_mblfs.Register(test_var_name, test_td_mblfs);
-  }
-}
-
-void
-TimeDependentEquationSystem::EliminateCoupledVariables()
-{
-  for (const auto & test_var_name : _test_var_names)
-  {
-    auto td_blf = _td_blfs.Get(test_var_name);
-    auto lf = _lfs.Get(test_var_name);
-    *lf *= _dt_coef.constant;
-    // if implicit, add contribution to linear form from terms involving state
-    // The AddMult method in mfem::BilinearForm is not defined for non-legacy assembly
-    mfem::Vector lf_prev(lf->Size());
-    td_blf->Mult(*_eliminated_variables.Get(test_var_name), lf_prev);
-    *lf += lf_prev;
-    for (const auto & eliminated_var_name : _eliminated_var_names)
-    {
-      if (_td_mblfs.Has(test_var_name) && _td_mblfs.Get(test_var_name)->Has(eliminated_var_name))
-      {
-        auto td_mblf = _td_mblfs.Get(test_var_name)->Get(eliminated_var_name);
-        // The AddMult method in mfem::BilinearForm is not defined for non-legacy assembly
-        mfem::Vector lf_prev(lf->Size());
-        td_mblf->Mult(*_eliminated_variables.Get(eliminated_var_name), lf_prev);
-        *lf += lf_prev;
-      }
-    }
-  }
-  // Eliminate contributions from other coupled variables.
-  EquationSystem::EliminateCoupledVariables();
-}
-
-void
-TimeDependentEquationSystem::UpdateEquationSystem()
-{
-  EquationSystem::BuildEquationSystem();
 }
 
 } // namespace Moose::MFEM
